api/smart_product_search/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
pydantic==2.11.7
spacy==3.8.7
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
typing_extensions==4.14.1
fastapi-cache2==0.2.2
